assets/model_monitoring/components/tests/requirements.txt
<<<<<<< HEAD
azure-ai-ml==1.12.1
=======
azure-ai-ml==1.13.0
>>>>>>> 7a54b91f3a492ed00e3033a99450bbc4df36a0fa
azure-cli-core~=2.45.0
azure-identity~=1.12.0
azureml-fsspec~=1.0.0
azureml-mlflow~=1.49.0
fsspec~=2023.4.0
mltable~=1.3.0
numpy~=1.19.0
pandas~=1.4.3
pyspark
pytest~=7.2.1
interpret-community[mimic]~=0.31.0
lightgbm~=4.3.0
ruamel.yaml==0.17.21
scipy~=1.10.0
pytest-xdist
pytest-split
pyspark-test~=0.2.0
azureml-metrics[tabular]~=0.0.42